Move StandardTrace class to separate file
