+++ aleth-vm/main.cpp
@@ -13,6 +13,7 @@
 #include <libethereum/ChainParams.h>
 #include <libethereum/Executive.h>
 #include <libethereum/LastBlockHashesFace.h>
+#include <libethereum/StandardTrace.h>
 #include <libevm/VMFactory.h>
 
 #include <aleth/buildinfo.h>
