+++ libweb3jsonrpc/Debug.cpp
@@ -1,11 +1,12 @@
+#include "Debug.h"
+#include "JsonHelper.h"
 #include <jsonrpccpp/common/exception.h>
 #include <libdevcore/CommonIO.h>
 #include <libdevcore/CommonJS.h>
 #include <libethcore/CommonJS.h>
 #include <libethereum/Client.h>
 #include <libethereum/Executive.h>
-#include "Debug.h"
-#include "JsonHelper.h"
+#include <libethereum/StandardTrace.h>
 using namespace std;
 using namespace dev;
 using namespace dev::rpc;
